Module graphics_lib::three_d

The three_d module provides support for loading and rendering 3D models.

Currently, the only supported file format for models is OBJ/MTL.

The engine can render objects using a variety of shaders, such as Blinn-Phong or Gouraud. The engine can also render objects using deferred Blinn-Phong rendering, allowing for a arbitrary number of lights. This module also provides a raytracing engine for rendering scenes

Modules

animation

A trait representing a time-dependent animation, along with a few implementations of simple use cases.

buffer

Wrapper structs around OpenGL Buffers.

lights

A module with structs representing different kinds of lights.

raytracing

This module provides support for a basic raytracing engine.

scene

A Scene represents a set of lights, shapes, and a specific rendering program.

shaders

A wrapper around an OpenGL shader program

shape

A shape with given vertex coordinates and normal vectors.

vao

This module provides a wrapper around an OpenGL VertexArrayObject