Module graphics_lib::three_d::animation

A trait representing a time-dependent animation, along with a few implementations of simple use cases.

Structs

Composite

This struct represents a combination of animations.

ConstantRotation

This represents a constant rotation with some given period along a specified axis.

Rotation

This represents a rotation along a specified axis, with the angle depending on some closure.

Scaling

This represents scaling in all three dimensions, according to specified closures.

Translation

This represents a translation in along all three axes, specified by closures.

Traits

Animation

An Animation is any struct that can define a transformation on an object. This could be rotating the object, or sliding the object for some time, etc.