Struct graphics_lib::three_d::animation::Scaling

pub struct Scaling {
    pub x_func: Option<fn(_: f32) -> f32>,
    pub y_func: Option<fn(_: f32) -> f32>,
    pub z_func: Option<fn(_: f32) -> f32>,
}

This represents scaling in all three dimensions, according to specified closures.

If None is given instead of a function, the value will stay the same was what it was the previous timestep.

Fields
x_func: Option<fn(_: f32) -> f32>

The scaling in the x-dimension

y_func: Option<fn(_: f32) -> f32>

The scaling in the y-dimension

z_func: Option<fn(_: f32) -> f32>

The scaling in the z-dimension