Module graphics_lib::three_d::buffer

Wrapper structs around OpenGL Buffers.

This module provides wrappers around Vertex and Normal buffers (VBOs), Index buffers (EBOs), and Frame buffers.

Structs

FrameBuffer

A wrapper around a frame buffer and its associated textures.

IndexBuffer

A wrapper about an index buffer (EBO).

NormalBuffer

A wrapper struct around a normal buffer (VBO).

VertexBuffer

A wrapper struct around a vertex buffer (VBO).

Type Definitions

Normal

A type alias for a normal, or a three-dimensional vector

Vertex

A type alias for a vertex, or a point in 3D space.