Module graphics_lib::three_d::raytracing

This module provides support for a basic raytracing engine.

The engine provides raytraced spheres and boxes, materials like metal and glass, and a camera to render them with.

It also provides traits for general materials and raytracing objects

Modules

aabb

An axis-aligned bounding box

bvh

A Bounding Volume Hierarchy

camera

A camera to render raytracing scenes

hit_record

A record of information about a ray intersection

interval

An interval along the number line

lights

A material for objects that emit light

material

A raytracing material

onb

An othornormal basis, used for transforming vectors relative to prespectives

opengl

Renders the pixel data

pdf

A Probability Density Function

ray

A ray, with an origin and direction

shape

A trait representing a raytracing object

vector

A 3-dimensional vector

Functions

random_cosine_direction

Generate a random vector in a random direction, without having to reject rays