Struct graphics_lib::three_d::raytracing::hit_record::HitRecord

pub struct HitRecord {
    pub p: Vec3,
    pub normal: Vec3,
    pub mat: Arc<dyn Material>,
    pub t: f32,
    pub front_face: bool,
}

A struct representing the result of a hit

Fields
p: Vec3

The point that was hit at (p = r0 + r*t)

normal: Vec3

The normal vector to the hit

mat: Arc<dyn Material>

The material that was hit

t: f32

The ‘t’ at which the ray hit (using the ray form r0 + r*t)

front_face: bool

Whether the ray hit the front or back face of the object

Implementations

impl HitRecord

pub fn set_face_normal(&mut self, r: &Ray, outward_normal: &Vec3)

Determine in which direction the normal points, and whether the ray hit the front of a face or the back

pub fn blank_with_mat(mat: Arc<dyn Material>) -> HitRecord

Create a default material, except that it has the given material

pub fn new( root: f32, r: &Ray, center: &Vec3, radius: f32, mat: Arc<dyn Material> ) -> HitRecord

Create a new material, given information about the circle it hit This method specifically focuses on circles, but this class is general purpose

pub fn self_without_mat(&self) -> HitRecordNoMat

Return the hit record without the material