Struct graphics_lib::three_d::raytracing::interval::Interval

pub struct Interval {
    pub min: f32,
    pub max: f32,
}

A struct representing an interval between two numbers

Fields
min: f32

The minimum value of the interval

max: f32

The maximum value of the interval

Implementations

impl Interval

pub fn empty() -> Interval

Create an empty interval, which contains no numbers

pub fn new(min: f32, max: f32) -> Interval

Create an interval, given two endpoints

pub fn new_from_intervals(a: Interval, b: Interval) -> Interval

Create a super-interval that encapsulates two intervals

pub fn contains(&self, x: f32) -> bool

Check if the interval contains a number

pub fn surrounds(&self, x: f32) -> bool

Check if the interval surrounds the number This method rejects values that are on the endpoints

pub fn clamp(&self, x: f32) -> f32

Clamp a value to be within the array

pub fn size(&self) -> f32

Compute the size of the interval

pub fn expand(&self, delta: f32) -> Interval

Expand the interval, adding half of delta on each side

pub fn replace_max(&mut self, max: f32) -> Interval

Replace the maximum value with a new value, and then return the interval