Trait graphics_lib::three_d::raytracing::material::Material

pub trait Material: Send + Sync {
    // Required method
    fn scatter(
        &self,
        ray_in: Ray,
        rec: HitRecordNoMat
    ) -> Option<(Vec3, Ray, Option<f32>)>;

    // Provided methods
    fn emitted(
        &self,
        _ray_in: Ray,
        _rec: HitRecordNoMat,
        _u: f32,
        _v: f32,
        _p: Vec3
    ) -> Vec3 { ... }
    fn scattering_pdf(
        &self,
        _ray_in: Ray,
        _rec: &HitRecord,
        _scattered: Ray
    ) -> f32 { ... }
}

A trait representing a generic material

Required Methods

fn scatter( &self, ray_in: Ray, rec: HitRecordNoMat ) -> Option<(Vec3, Ray, Option<f32>)>

Compute how a ray would scatter off of the material, given information about the ray intersection If the ray scatters, return information about the color of the scattered ray, the scattering direction, and the {DF (if supported) If the ray doesn’t, return None

Provided Methods

fn emitted( &self, _ray_in: Ray, _rec: HitRecordNoMat, _u: f32, _v: f32, _p: Vec3 ) -> Vec3

Compute the emitted light of a material /// This method is optional

fn scattering_pdf(&self, _ray_in: Ray, _rec: &HitRecord, _scattered: Ray) -> f32

Compute the PDF of the material This method is optional