Struct graphics_lib::three_d::raytracing::onb::ONB

pub struct ONB { /* private fields */ }

A struct representing an orthonormal basis An orthonormal basis is a collection of three vectors that are are perpendicular to each other

Implementations

impl ONB

pub fn u(&self) -> Vec3

Get the first vector

pub fn v(&self) -> Vec3

Get the second vector

pub fn w(&self) -> Vec3

Get the third vector

pub fn local(&self, a: f32, b: f32, c: f32) -> Vec3

Transform the input vector (as component values) to be relative to the ONB

pub fn local_from_vector(&self, a: Vec3) -> Vec3

Transform the input vector to be relative to the ONB

pub fn build_from_w(w: Vec3) -> ONB

Create an ONB from a given vector