Trait graphics_lib::three_d::raytracing::shape::RTObject

pub trait RTObject: Send + Sync {
    // Required methods
    fn ray_intersects(&self, r: &Ray, ray_t: Interval) -> Option<HitRecord>;
    fn bounding_box(&self) -> AABB;
    fn clone_dyn(&self) -> Box<dyn RTObject>;

    // Provided methods
    fn pdf_value(&self, o: Vec3, v: Vec3) -> f32 { ... }
    fn random(&self, o: Vec3) -> Vec3 { ... }
}

A trait representing a raytracing object A raytracing object is a 3-dimensional object that provides methods for computing ray bounces and lighting

Required Methods

fn ray_intersects(&self, r: &Ray, ray_t: Interval) -> Option<HitRecord>

Compute whether the given ray in the specified interval intersects with the object If the ray intersects the object, this method should return a HitRecord with information about the intersection Otherwise, this value should return None

fn bounding_box(&self) -> AABB

Compute a bounding box that contains the object

fn clone_dyn(&self) -> Box<dyn RTObject>

Clone the object into a Box

Provided Methods

fn pdf_value(&self, o: Vec3, v: Vec3) -> f32

Compute the probability that a ray starting from a certain point with a certain direction would be reflected by this object This method is not required

fn random(&self, o: Vec3) -> Vec3

Compute a random ray bouncing off the object, given the origin of the vector This should take into account the geometry of the object to match the real probabilities of reflecting in different directions This method is not required