Struct graphics_lib::three_d::raytracing::vector::Vec3

pub struct Vec3 {
    pub data: [f32; 3],
}

A struct representing a three-dimensional vector

Fields
data: [f32; 3]

The x, y, and z components of the vector

Implementations

impl Vec3

pub fn new(data: [f32; 3]) -> Vec3

Create a new vector from the given array

pub fn length(&self) -> f32

Compute the length of the vector

pub fn unit(&self) -> Vec3

Get the unit vector with the same direction as the input vector

pub fn x(&self) -> f32

Get the value of the x-component of the vector

pub fn y(&self) -> f32

Get the value of the y-component of the vector

pub fn z(&self) -> f32

Get the value of the z-component of the vector

pub fn for_each(&mut self, func: &fn(_: f32) -> f32)

Apply a function to each of the components of the vector

pub fn length_squared(&self) -> f32

Compute the squared length of the vector

pub fn near_zero(&self) -> bool

Check whether all the values of the vector are within 1e-8 of 0

impl Vec3

pub fn dot(v1: &Vec3, v2: &Vec3) -> f32

Compute the dot product of two vectors

pub fn cross(v1: &Vec3, v2: &Vec3) -> Vec3

Compute the cross product of two vectors

pub fn random(rng: &mut ThreadRng) -> Vec3

Generate a vector of random values

pub fn random_in_range<R>(rng: &mut ThreadRng, range: R) -> Vec3where R: SampleRange<f32> + Clone,

Generate a vector with random values within the given range

pub fn random_in_unit_sphere(rng: &mut ThreadRng) -> Vec3

Generate a random vector within the unit circle This vector is not necessarily on the unit circle, just inside it

pub fn random_on_hemisphere(rng: &mut ThreadRng, normal: &Vec3) -> Vec3

Generate a random vector in a hemisphere in the direction of the normal vector

pub fn reflect(v: &Vec3, n: &Vec3) -> Vec3

Reflect off of a surface, given the incoming vector and the normal vector of the surface

pub fn refract(uv: &Vec3, n: &Vec3, etai_over_etat: f32) -> Vec3

Refract through a material, given the incoming vector, the normal vector of the surface, and the refraction ratio