Module graphics_lib::three_d::shaders

A wrapper around an OpenGL shader program

Structs

BLINN_PHONG

Basic Blinn-Phong shading.

BLINN_PHONG_LIGHTING

Program for the lighting stage of deferred rendering, using Blinn-Phong shading.

BLINN_PHONG_POINT_LIGHTING

Program for the point lighting stage of deferred rendering, using Blinn-Phong shading.

EMISSION

Program for adding emission lighting to deferred rendering.

PREPASS

Program for the prepass step of deferred rendering.

Program

A wrapper struct around an OpenGL program.

Shader

A wrapper struct around an OpenGL shader.

Enums

ShaderProgramType

Represents the types of shader.

ShaderType

This enum represents the default shader types.

Constants

BLINN_PHONG_3D_FRAG_SHADER

Shader code for a Blinn-Phong fragment shader.

BLINN_PHONG_3D_LIGHTING_FRAG_SHADER

Blinn-Phong fragment shader code for the lighting step of deferred rendering.

BLINN_PHONG_3D_POINT_LIGHTING_FRAG_SHADER

Blinn-Phong fragment shader code for the point lighting step of deferred rendering.

BLINN_PHONG_3D_SHADER

Shader code for a Blinn-Phong shader.

DEFAULT_3D_FRAG_SHADER

Shader code for a basic diffuse fragment shader.

DEFAULT_3D_SHADER

Shader code for a basic diffuse shader.

EMISSION_FRAG_SHADER

deferred rendering shader that adds emission lighting.

GOURAUD_3D_FRAG_SHADER

Shader code for a Gouraud fragment shader.

GOURAUD_3D_SHADER

Shader code for a Gouraud shader.

LIGHTING_SHADER

Blinn-Phong shader code for the lighting step of deferred rendering.

PREPASS_FRAG_SHADER

Fragment shader code for the prepass step of deferred rendering.

PREPASS_SHADER

Shader code for the prepass step of deferred rendering.

Functions

compile_shader

Compiles a shader, given the shader type and the source code.

link_program

Links a vertex shader and a fragment shader into a program.