Module graphics_lib::three_d::shape::importing

Methods for loading a shape from modeling files.

Structs

Material

This struct represents an MTL material/

Enums

IlluminationModel

These are all the illumination models of the OBJ file format. An illumination model roughly defines what type of rendering program should be used to render the object.

ImportError

An enum represents errors that can happen in the importing process.

Constants

DEFAULT_SPEC_EXP

If an object has no specified specular exponent, this value is used.

MAT

FIXME: Remove this!