Struct graphics_lib::three_d::shape::Shape

pub struct Shape { /* private fields */ }

A shape with given vertex coordinates and normal vectors.

Implementations

impl Shape

pub fn from_obj(
path: &str,
shader_type: ShaderType,
transform: Option<Transform>,
animation: Option<Box<dyn Animation>>,
_bface_culling: bool
) -> Result<Shape, ImportError>

Loads the shape from the given obj file.

Even if the obj file defines objects or vertex groups, it will all be imported as one shape. This is because there is no good way to find the origins of the objects/groups, so transformations of those shapes would be incorrect. Additionally, there are good reasons why a modeler would want multiple objects to be packed into the same shape. If you want to load two different shapes, split the shapes across files.

impl Shape

pub fn bind_attributes(&self, program: &Program)

Bind the position and normal attributes of this shape to the given program.

impl Shape

pub fn animate(&mut self, t: f32)

Run one step of animation, which updates the shape’s 3D transformation based on the shape’s animation.

pub fn replace_animation(&mut self, animation: Box<dyn Animation>)

Swap out the current animation for a different one.

pub fn set_material(&mut self, mat: Material)

Sets the shape’s material.

impl Shape

pub fn set_transform_matrix(
&mut self,
scaling: Option<Mat4>,
rotation: Option<Mat4>,
translation: Option<Mat4>
)

Sets the shape’s transformation matrix.

pub fn set_scaling(&mut self, scaling: Mat4)

Sets the shape’s scaling matrix.

pub fn set_rotation(&mut self, rotation: Mat4)

Sets the shape’s rotation matrix.

pub fn set_translation(&mut self, translation: Mat4)

Sets the shape’s translation matrix.

impl Shape

pub fn draw(
&self,
light: &DirectionLight,
view: &Mat4,
program: &Program,
dims: (f32, f32)
)

Draws the shape to the set framebuffer, given the light source, the view matrix, the program to use, and the screen dimensions.